feat(file-manager): update get user logic
