+++ handlers/rootHandler.js
@@ -1,12 +1,15 @@
+import { cwd } from 'node:process'
 import { handleUserOperationSystem } from './operatingSystemHandlers.js'
 import { handleUserBasicOperation } from './basicOperationHandlers.js'
 import { handleNavigationOperation } from './navigationOperationHandlers.js'
 import { calculateHash } from './hashOperationHandlers.js'
 import { handleZipOperation } from './zipOperationHandlers.js'
+import { printErrorText } from '../utils/colorTextUtils.js'
 import { BASIC_OPERATIONS, NAVIGATION_OPERATIONS, ZIP_OPERATIONS } from '../utils/constants.js'
 
-export const handleUserInput = async (userInput, currentDirectory, readLine) => {
+export const handleUserInput = async (userInput, readLine) => {
   const [operationType, ...args] = userInput.trim().split(/\s+/g)
+  const currentDirectory = cwd()
 
   try {
     switch (operationType) {
