+++ index.js
@@ -1,43 +1,18 @@
-import { argv } from 'node:process'
 import readlinePromises from 'node:readline/promises'
-import { colorText } from './utils/colorTextUtils.js'
 import { printCurrentWorkingDirectory } from './utils/pathUtils.js'
+import { initAppHandler } from './handlers/initAppHandler.js'
 import { handleUserInput } from './handlers/rootHandler.js'
 
-// ToDO delete it after move logic
-import { fileURLToPath } from 'url'
-import { dirname } from 'path'
-
-const getUserName = (commandLineArguments) => {
-  const [ userNameArgument ] = commandLineArguments.splice(2)
-  const userName = userNameArgument && userNameArgument.replace('--', '')
-  return userName
-}
-
 const initFileManagerApp = async () => {
-  const userName = getUserName(argv)
-
-// ToDO delete it after move logic
-  const filePath = fileURLToPath(import.meta.url)
-  const dirPath = dirname(filePath)
-
-  if (!userName) {
-    const errorMessage = `Invalid username argument.\n Please start File Manager with command: "npm run start -- --username=your_username" \n` 
-    throw new Error(errorMessage)
-  }
-
-  console.log(colorText(`Welcome to the File Manager, ${userName}! \n`, 'green'))
-  printCurrentWorkingDirectory()
+  const userName = initAppHandler()
 
   const rl = readlinePromises.createInterface({
     input: process.stdin,
     output: process.stdout,
   })
 
-  process.chdir(dirPath)
-
   rl.on(`line`, async (line) => {
-    await handleUserInput(line, dirPath, rl)
+    await handleUserInput(line, rl)
     printCurrentWorkingDirectory()
   })
 
@@ -45,7 +20,7 @@ const initFileManagerApp = async () => {
     console.log(`Thank you for using File Manager, ${userName}, goodbye!\n`)
   })
 
-  rl.on('error', () => console.log('By by my darling'))
+  rl.on('error', () => console.log('An unknown error has occurred'))
 }
 
 await initFileManagerApp()
